refactor(cli): isolate wizard presentation
